% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anomes_to_date.R
\name{anomes_to_date}
\alias{anomes_to_date}
\title{Title}
\usage{
anomes_to_date(dat, first_day = 1)
}
\arguments{
\item{first_day}{numeric \code{vector} containing the day number(s) that will be used when
building the \code{date} object.}

\item{x}{}
}
\description{
Title
}
\details{
For February \code{first_day} must be 29 in non-leap years, and 28 on leap years.
}
