% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basin_names.R
\name{basin_names}
\alias{basin_names}
\title{Names of Spanish basins for the CEDEX gauging station data}
\usage{
basin_names(basin_nam = NULL)
}
\arguments{
\item{basin_nam}{\code{character} with the name of the basins to retrieve
\code{table_name} names for. Default is to retrieve data for all basins on the \emph{CEDEX} web site.}
}
\value{
A \code{data.frame} with a column named "name" that contains the names of the basins, and a second
column named "nameceh" with the basin names required for the web site.
}
\description{
\code{basin_names} retrieves basin names where gauging stations belonging to the CEDEX network
can be found.#'
}
\examples{
df <- basin_names(c("duero", "EbrO"))
}
