% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_url_files.R
\name{check_url_files}
\alias{check_url_files}
\title{Check URLs and return full http addresses.}
\usage{
check_url_files(url, basin_names, file_name, sf, verbose)
}
\arguments{
\item{url}{a valid URL.}

\item{basin_names}{its correctness is not checked.}

\item{file_name}{}

\item{sf}{}

\item{verbose}{}
}
\value{
Full url for files.
}
\description{
Check URLs and return full http addresses.
}
\keyword{internal}
