% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correction_coordinates.R
\name{correction_coordinates}
\alias{correction_coordinates}
\title{Correction of lat-long coordinates}
\usage{
correction_coordinates(df)
}
\arguments{
\item{df}{data.frame with a column named "indroea" and two other columns
for latitude and longitude.}
}
\value{
The input data.frame with corrected coordinates. If column "indroea" is not present
\code{correction_coordinates} returns the input \code{data.frame}.
}
\description{
A correction is introduced for some lat-long coordinates in some basins.
}
\details{
A correction to coordinates, email from Carmen Mirta Dimas Suárez (CEH)
on July 17th, 2024. The correction applies to stations 3179, 3191, 3275 and 5107 only.
}
\examples{
df <- data.frame(indroea = c("3179", "3000"), lat = 1:2, long = 3:4)
correction_coordinates(df)

}
\keyword{internal}
