% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epsg_ceh.R
\name{epsg_ceh}
\alias{epsg_ceh}
\title{EPSG codes for CEH coordinates}
\usage{
epsg_ceh(x)
}
\arguments{
\item{x}{character vector with a short string description of the coordinate system.
It can contain the string "utm30", "etrs89", "wgs84" or "ed50". Upper/Lower case letters are
also accepted.}
}
\value{
Corresponding EPSG codes in a named vector.
}
\description{
It translates the "utm30", "etrs89", "wgs84" or "ed50" strings into EPSG codes as used in the coordinate
systems at the gauging stations of the CEH.
}
\details{
Simple implementation of \code{switch}.
}
\examples{
epsg_ceh(c("ed50", "wgs84"))
}
