% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full_name_url.R
\name{full_name_url}
\alias{full_name_url}
\title{Return full URL addresses.}
\usage{
full_name_url(url, file_name, basin_nam, sf = TRUE)
}
\arguments{
\item{url}{a valid URL.}

\item{file_name}{\code{data.frame} of one row and three columns named 'file', 'file_coords' and 'id_join',
\emph{CEDEX} site, without extension.}

\item{basin_nam}{\code{character} with the name of the basins to retrieve
\code{table_name} names for.}

\item{sf}{logical, if TRUE (default),}
}
\value{
\code{data.frame} with one column named 'files' containing the full url address for files.
if \code{sf} is TRUE, a second column named 'coord' is added that includes the full url address
of the files where coordinates will be found.
}
\description{
\code{full_name_url} builds full URL addresses
}
\details{
'file_name' will be pasted to the input 'URL' on output.
}
\examples{
url <- "https://ceh-flumen64.cedex.es/anuarioaforos//anuario-2020-2021/"
file_name <- data.frame(file = "afliq", file_coords = "estaf", id_join = "indroea")
x <- full_name_url(url, file_name, "ebro")

}
\keyword{internal}
