% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ceh_data.R
\name{get_ceh_data}
\alias{get_ceh_data}
\title{Read Spanish basin data from the \emph{cedex} site.}
\usage{
get_ceh_data(
  table = "estaf",
  basin = "all",
  sf = TRUE,
  crs = "utm30",
  verbose = TRUE
)
}
\arguments{
\item{table}{\code{character} with the name of the file to retrieve from the
\emph{cedex} site, without extension. If not given, the default value is "estaf".}

\item{basin}{}

\item{sf}{logical, if TRUE \code{read_ceh_basin} returns a \code{sf} spatial object.}

\item{crs}{\code{character} indicating the coordinate system of choice. Values can be
"UTM", "UTM30" (default), "WGS84", "ED50" or "ETRS89" in lower or upper case letters (e.g. "wgs84" is valid,
as is "Wgs84").}

\item{verbose}{\code{logical}, if set to TRUE progress messages are printed on screen.}
}
\value{
A spatial \code{sf} object with a EPSG coordinate reference system (unless cs = "utm").
}
\description{
\code{read_ceh_basin} retrieves time series datasets and statistics for
several Spanish basins from the Centro de Estudios y Experimentación
de Obras Públicas (CEDEX).
}
\details{
To see a description of the files to retrieve see
"https://ceh.cedex.es/anuarioaforos/demarcaciones.asp" and go to the basin you want
to get data for.
Regarding coordinates, crs = "UTM" indicates local UTM coordinates, which may be a problem
if the sites are located on both sides of the UTM demarcation line, so we do not recommend it.
crs = "UTM30" actually means Mercator UTM 30 North with datum WGS84, whereas ETRS89 implies
Mercator UTM 30 with datum ETRS89. Finally, crs = "ED50" and "WGS84" are longitude-latitude
system with those datums.

The \code{encoding} argument of the \code{read.csv2} function is set to "latin1" in order to
correctly read diacritics and tildes.
}
