% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_leap_year.R
\name{is_leap_year}
\alias{is_leap_year}
\title{Check whether year is leap}
\usage{
is_leap_year(year)
}
\arguments{
\item{year}{\code{integer} vector containing (positive) years.}
}
\value{
Logical \code{TRUE} or \code{FALSE} vector depending on whether year is leap or not.
}
\description{
\code{is_leap_year} checks whether numbers in input vector correspond to leap years.
}
\details{
Straightforward algorithm. Input \code{year} must contain numbers without decimals,
although its \code{typeof} value can be \code{"integer"} or \code{"double"}.
}
\examples{
# Check leap years for XX century.
is_leap_year(seq(1900, 1999))

}
