% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cedex_basin.R
\name{read_cedex_basin}
\alias{read_cedex_basin}
\title{Read Spanish basin data from the \emph{cedex} site.}
\usage{
read_cedex_basin(
  table = "estaf",
  basin = "all",
  sf = TRUE,
  cs = "utm30",
  url = NULL,
  verbose = TRUE
)
}
\arguments{
\item{table}{\code{character} with the name of the file to retrieve from the
\emph{cedex} site, without extension. If not given, the default value is "estaf".}

\item{cs}{\code{character} indicating the coordinate system of choice. Values can be
"UTM", "UTM30" (default), "WGS84", "ED50" or "ETRS89" in lower or upper case letters (e.g. "wgs84" is valid,
as is "Wgs84").}

\item{verbose}{\code{logical}, if set to TRUE progress messages are printed on screen.}

\item{str_url}{string with the \code{url} of the \emph{cedex} site. Value by default is
"https://ceh-flumen64.cedex.es/anuarioaforos//anuario-2019-2020//EBRO//", corresponding to the
Ebro basin.
Notice that, internally, \code{read_cedex_basin} will paste \code{table} and \code{str_url} together,
so any missing slash at the end of \code{str_url} will raise an error.}
}
\value{
A spatial \code{sf} object with a EPSG coordinate reference system (unless cs = "utm").
}
\description{
\code{read_cedex_basin} retrieves time series datasets and statistics for
several Spanish basins from the Centro de Estudios y Experimentación
de Obras Públicas (CEDEX).
}
\details{
To see a description of the files to retrieve see
"https://ceh.cedex.es/anuarioaforos/demarcaciones.asp" and go to the basin you want
to get data for.
Regarding coordinates, cs = "UTM" indicates local UTM coordinates, which may be a problem
if the sites are located on both sides of the UTM demarcation line, so we do not recommend it.
cs = "UTM30" actually means Mercator UTM 30 North with datum WGS84, whereas ETRS89 implies
Mercator UTM 30 with datum ETRS89. Finally, cs = "ED50" and "WGS84" are longitude-latitude
system with those datums.

The \code{encoding} argument of the \code{read.csv2} function is set to "latin1" in order to
correctly read diacritics and tildes.
}
