% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_basins.R
\name{select_basins}
\alias{select_basins}
\title{Names of Spanish basins for the CEDEX gauging station data}
\usage{
select_basins(basin_nam = NULL)
}
\arguments{
\item{basin_nam}{\code{character} with the name of the basins to retrieve
\code{table_name} names for. Default is to retrieve data for all basins on the \emph{CEDEX} web site.}
}
\value{
A \code{data.frame} with a column named "name" that contains the names of the basins, and a second
column named "nameceh" with the basin names required for the web site.
}
\description{
\code{select_basins} retrieves basin names where gauging stations belonging to the CEDEX network
can be found.#'
}
\details{
It uses internal package data, so it will only work when used by other package functions.
}
\examples{
df <- select_basins(c("duero", "EbrO"))

}
