% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_coordinates.R
\name{select_coordinates}
\alias{select_coordinates}
\title{Select where to find coordinates for CEDEX file}
\usage{
select_coordinates(table_name = "estaf")
}
\arguments{
\item{table_name}{\code{character} with the name of the file to retrieve from the
\emph{CEDEX} site, without extension. If not given, the default value is "estaf".}
}
\value{
A \code{data.frame} of one row only with three fields: "file", containing the value
of the "table_name" input, "file_coords", indicating the name of the CEDEX file where
coordinates will be found, and "id_join", mentioning which key can be used for left-joins.
}
\description{
\code{select_coordinates} indicates where to find spatial coordinates for \code{table_name}.
}
\examples{
x <- select_coordinates("afliq")

}
\keyword{internal}
