% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_url_files.R
\name{check_url_files}
\alias{check_url_files}
\title{Check URLs and return full http addresses.}
\usage{
check_url_files(url, basin, table, verbose)
}
\arguments{
\item{url}{a valid URL.}

\item{basin}{a character vector containing the names of the basins to download the data for.}

\item{table}{a character string indicating which table to download.}
}
\value{
Full url for files.
}
\description{
Check URLs and return full http addresses.
}
\keyword{internal}
