% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_url_files.R
\name{check_url_files}
\alias{check_url_files}
\title{Check URLs and return full http addresses.}
\usage{
check_url_files(
  url,
  file_name,
  basin_nam,
  timeout = 120,
  sf = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{url}{a valid URL.}

\item{file_name}{\code{data.frame} of one row and three columns named 'file', 'file_coords' and 'id_join',
\emph{CEDEX} site, without extension.}

\item{basin_nam}{\code{character} with the name of the basins to retrieve
\code{table_name} names for.}

\item{timeout}{positive integer specifying the timeout for some Internet operations, in seconds.
Default is 120 seconds. Depending on the bandwidth of your internet connection or on the state of the
CEDEX servers you may have to set a \code{timeout} value longer than 120.}

\item{sf}{logical, if TRUE (default), \code{hydrospain} returns a \code{sf} spatial object.
Coordinate system is always \code{EPSG:32630}, which corresponds to WGS84 / UTM zone 30N.}

\item{verbose}{\code{logical}, if set to TRUE progress bars are printed on screen.}
}
\value{
Full url for files.
}
\description{
\code{check_url_files} verifies that the requests to 'file_name' files are ok.
}
\details{
'file_name' will be pasted to the input 'URL' and a status request will be sent.
}
\examples{

url <- "https://ceh-flumen64.cedex.es/anuarioaforos//anuario-2020-2021/"
file_name <- data.frame(file = "afliq", file_coords = "estaf", id_join = "indroea")

\donttest{
x <- check_url_files(url, file_name, "ebro", timeout = 300)
}

}
\keyword{internal}
