% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrospain.R
\name{hydrospain}
\alias{hydrospain}
\title{Read data from gauging stations of most largest Spanish rivers}
\usage{
hydrospain(
  file_name = "estaf",
  basin_nam = NULL,
  timeout = 120,
  first_day = 1,
  sf = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{file_name}{\code{character} string with the name of the file to retrieve from the
\emph{CEDEX} site, without extension. If not given, the default value is "estaf". More than one name is
not allowed.}

\item{basin_nam}{\code{character} with the name of the basins to retrieve
\code{table_name} names for. Default is to retrieve data for all basins on the \emph{CEDEX} web site.}

\item{timeout}{positive integer specifying the timeout for some Internet operations, in seconds.
Default is 120 seconds. Depending on the bandwidth of your internet connection or on the state of the
CEDEX servers you may have to set a \code{timeout} value longer than 120.}

\item{first_day}{numeric or character \code{vector} containing the day number(s) that will be set when
building the \code{date} object.}

\item{sf}{logical, if TRUE (default), \code{hydrospain} returns a \code{sf} spatial object.
Coordinate system is always \code{EPSG:32630}, which corresponds to WGS84 / UTM zone 30N.}

\item{verbose}{\code{logical}, if set to TRUE progress bars are printed on screen.}
}
\value{
A spatial \code{sf} object with a WGS84/UTM zone 30N coordinate reference system (EPSG:32630).
}
\description{
\code{hydrospain} retrieves time series datasets and statistics for
several Spanish basins from the Centro de Estudios Hidrológicos (CEX) of
the Centro de Estudios y Experimentación de Obras Públicas (CEDEX).
}
\details{
To see a description of the files to retrieve see
"https://ceh.cedex.es/anuarioaforos/demarcaciones.asp" and go to the basin you want
to get data for.

The coordinate reference system of the resulting \code{sf} object is \code{UTM 30N} always.
There are coordinates in other systems in the original CEDEX files but they are not used
(although they are also retrieved and included in the output object). Notice that long-lat coordinates
are given as 'DDMMSS' strings only. Future versions will provide those coordinates in numeric format.
}
\examples{
\dontrun{
# Read afliq.csv data from basin 'cantabrico'.
x <- hydrospain(file_name = "afliq", basin_nam = "cantabrico", verbose = FALSE)
}

}
